refactor(ops): Have cargo-metadata's ops match the command name
