+++ src/ops/cargo_metadata.rs
